// bridge/typechain-types/StakingDescriptorMock.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import { FunctionFragment, Result } from "@ethersproject/abi";
import { Listener, Provider } from "@ethersproject/providers";
import {
  BaseContract,
  BigNumber,
  BigNumberish,
  BytesLike,
  CallOverrides,
  PopulatedTransaction,
  Signer,
  utils,
} from "ethers";
import { OnEvent, TypedEvent, TypedEventFilter, TypedListener } from "./common";

export declare namespace StakingDescriptor {
  export type ConstructTokenURIParamsStruct = {
    tokenId: BigNumberish;
    shares: BigNumberish;
    freeAfter: BigNumberish;
    withdrawFreeAfter: BigNumberish;
    accumulatorEth: BigNumberish;
    accumulatorToken: BigNumberish;
  };

  export type ConstructTokenURIParamsStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    tokenId: BigNumber;
    shares: BigNumber;
    freeAfter: BigNumber;
    withdrawFreeAfter: BigNumber;
    accumulatorEth: BigNumber;
    accumulatorToken: BigNumber;
  };
}

export interface StakingDescriptorMockInterface extends utils.Interface {
  contractName: "StakingDescriptorMock";
  functions: {
    "constructTokenURI((uint256,uint256,uint256,uint256,uint256,uint256))": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "constructTokenURI",
    values: [StakingDescriptor.ConstructTokenURIParamsStruct]
  ): string;

  decodeFunctionResult(
    functionFragment: "constructTokenURI",
    data: BytesLike
  ): Result;

  events: {};
}

export interface StakingDescriptorMock extends BaseContract {
  contractName: "StakingDescriptorMock";
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: StakingDescriptorMockInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    constructTokenURI(
      params: StakingDescriptor.ConstructTokenURIParamsStruct,
      overrides?: CallOverrides
    ): Promise<[string]>;
  };

  constructTokenURI(
    params: StakingDescriptor.ConstructTokenURIParamsStruct,
    overrides?: CallOverrides
  ): Promise<string>;

  callStatic: {
    constructTokenURI(
      params: StakingDescriptor.ConstructTokenURIParamsStruct,
      overrides?: CallOverrides
    ): Promise<string>;
  };

  filters: {};

  estimateGas: {
    constructTokenURI(
      params: StakingDescriptor.ConstructTokenURIParamsStruct,
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    constructTokenURI(
      params: StakingDescriptor.ConstructTokenURIParamsStruct,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}